+++ agent-monolith/src/main/java/xwsagent/wroomagent/soap/xsd/CommentSoap.java
@@ -71,7 +71,7 @@ public class CommentSoap {
     protected boolean deleted;
     protected boolean approved;
     protected int rate;
-    protected Integer replyId;
+    protected Long replyId;
     protected Boolean reply;
     protected long adId;
     protected Long clientId;
@@ -222,7 +222,7 @@ public class CommentSoap {
      *     {@link Integer }
      *     
      */
-    public Integer getReplyId() {
+    public Long getReplyId() {
         return replyId;
     }
 
@@ -234,7 +234,7 @@ public class CommentSoap {
      *     {@link Integer }
      *     
      */
-    public void setReplyId(Integer value) {
+    public void setReplyId(Long value) {
         this.replyId = value;
     }
